Create README.md
Cannabis Asset Units protocol is a combination of SafeToken structure with a PAXG type asset-backing and control mechanisms. The Dev will mint 1,000,000 tokens in the first 6 months to kickstart the project, and as the interim operator of the protocol will strive to mint 1,000,000 more biannually for 5 years. An initial ultra-low cap presale has been conducted on dxSale with a 4 year liquidity lock to demonstrate our intention for longevity. #Cannaconomy is the place where you can prosper with us.